+++ packages/keyring-eth-ledger-bridge/CHANGELOG.md
@@ -11,8 +11,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
-- Bump `@metamask/keyring-api` from `^23.5.0` to `^23.6.0` ([#9999](https://github.com/MetaMask/accounts/pull/9999))
-- Bump `@metamask/keyring-sdk` from `^2.2.0` to `^2.2.1` ([#9999](https://github.com/MetaMask/accounts/pull/9999))
+- Bump `@metamask/keyring-api` from `^23.5.0` to `^23.6.0` ([#1023](https://github.com/MetaMask/accounts/pull/1023))
+- Bump `@metamask/keyring-sdk` from `^2.2.0` to `^2.2.1` ([#1023](https://github.com/MetaMask/accounts/pull/1023))
 
 ## [12.3.0]